removed redundant tests, moved file onto test folder, and cleaned files for linting
