+++ backend/api/tests/test_views.py
@@ -7,13 +7,14 @@ import datetime
 import requests_mock
 import os
 from oauth2_provider.models import get_application_model, get_access_token_model
+
 Application = get_application_model()
 AccessToken = get_access_token_model()
 
 
 def setupOAuth(self):
-    CLIENT_ID = os.getenv('CLIENT_ID', 'defaultTestClientID')
-    CLIENT_SECRET = os.getenv('CLIENT_SECRET', 'defaultTestClientSecret')
+    CLIENT_ID = os.getenv("CLIENT_ID", "defaultTestClientID")
+    CLIENT_SECRET = os.getenv("CLIENT_SECRET", "defaultTestClientSecret")
     self.user = User.objects.create_superuser(
         email="superuserOAuth@email.com", password="superuser"
     )
@@ -24,7 +25,7 @@ def setupOAuth(self):
         client_type=Application.CLIENT_CONFIDENTIAL,
         name="OAuth-Test-API",
         redirect_uris="http://127.0.0.1:8000/authredirect/",
-        skip_authorization=True
+        skip_authorization=True,
     )
     self.application.save()
     self.token = AccessToken.objects.create(
@@ -40,19 +41,13 @@ def setupOAuth(self):
 class Test_UserView(APITestCase):
     def setUp(self):
         self.store1 = Store.objects.create(
-            address="1 Main Street",
-            name="Store 1",
-            category=Category.FOOD
+            address="1 Main Street", name="Store 1", category=Category.FOOD
         )
         self.store2 = Store.objects.create(
-            address="2 Main Street",
-            name="Store 2",
-            category=Category.FOOD
+            address="2 Main Street", name="Store 2", category=Category.FOOD
         )
         self.store3 = Store.objects.create(
-            address="3 Main Street",
-            name="Store 3",
-            category=Category.FOOD
+            address="3 Main Street", name="Store 3", category=Category.FOOD
         )
         self.userExists = User.objects.create_user(
             email="exists@example.com", password="password"
@@ -66,20 +61,14 @@ class Test_UserView(APITestCase):
 
     def test_register_good(self):
         url = "http://127.0.0.1:8000/api/register/"
-        user = {
-            "email": "testuser@example.com",
-            "password": "testpassword"
-        }
+        user = {"email": "testuser@example.com", "password": "testpassword"}
         r = self.client.post(url, user)
         self.assertEqual(201, r.status_code)
         self.assertEqual(1, r.data["status"])
 
     def test_register_bad(self):
         url = "http://127.0.0.1:8000/api/register/"
-        user = {
-            "email": "exists@example.com",
-            "password": "testpassword"
-        }
+        user = {"email": "exists@example.com", "password": "testpassword"}
         r = self.client.post(url, user)
         self.assertEqual(400, r.status_code)
         self.assertEqual(0, r.data["status"])
@@ -87,18 +76,15 @@ class Test_UserView(APITestCase):
     def test_retrieve_user(self):
         url = "http://127.0.0.1:8000/api/users/"
         r = self.client.get(
-            url + str(self.userExists.pk) + '/',
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
+            url + str(self.userExists.pk) + "/",
+            HTTP_AUTHORIZATION="Bearer " + self.token.token,
         )
         self.assertEqual(200, r.status_code)
         self.assertEqual("exists@example.com", r.data["email"])
 
     def test_list_user(self):
         url = "http://127.0.0.1:8000/api/users/"
-        r = self.client.get(
-            url,
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
-        )
+        r = self.client.get(url, HTTP_AUTHORIZATION="Bearer " + self.token.token)
         self.assertEqual(200, r.status_code)
         self.assertLessEqual(1, len(r.data))
 
@@ -117,11 +103,13 @@ class Test_UserView(APITestCase):
         pass
 
     def test_add_store(self):
-        url = "http://127.0.0.1:8000/api/users/" + str(self.userExists.pk) + "/add_store/"
+        url = (
+            "http://127.0.0.1:8000/api/users/" + str(self.userExists.pk) + "/add_store/"
+        )
         r = self.client.post(
             url,
             {"store_id": self.store3.pk},
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
+            HTTP_AUTHORIZATION="Bearer " + self.token.token,
         )
         self.assertEqual(200, r.status_code)
         self.assertEqual(2, len(self.userExists.stores.all()))
@@ -129,11 +117,15 @@ class Test_UserView(APITestCase):
         self.assertEqual(1, len(self.userExists.stores.all()))
 
     def test_remove_store(self):
-        url = "http://127.0.0.1:8000/api/users/" + str(self.userExists.pk) + "/remove_store/"
+        url = (
+            "http://127.0.0.1:8000/api/users/"
+            + str(self.userExists.pk)
+            + "/remove_store/"
+        )
         r = self.client.post(
             url,
             {"store_id": self.store2.pk},
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
+            HTTP_AUTHORIZATION="Bearer " + self.token.token,
         )
         self.assertEqual(200, r.status_code)
         self.assertEqual(0, len(self.userExists.stores.all()))
@@ -145,7 +137,7 @@ class Test_UserView(APITestCase):
         r = self.client.delete(
             url,
             {"user_id": self.userExists.pk, "store_id": self.store1.pk},
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
+            HTTP_AUTHORIZATION="Bearer " + self.token.token,
         )
         self.assertEqual(200, r.status_code)
         self.assertEqual(0, Store.objects.filter(name="Store 1").count())
@@ -179,14 +171,10 @@ class Test_StoreView(APITestCase):
                 description="Item 3",
             )
         self.store1 = Store.objects.create(
-            address="1 Main Street",
-            name="Store 1",
-            category=Category.FOOD
+            address="1 Main Street", name="Store 1", category=Category.FOOD
         )
         self.store2 = Store.objects.create(
-            address="2 Main Street",
-            name="Store 2",
-            category=Category.FOOD
+            address="2 Main Street", name="Store 2", category=Category.FOOD
         )
         self.store1.items.add(self.item2)
         self.store1.save()
@@ -194,18 +182,15 @@ class Test_StoreView(APITestCase):
     def test_retrieve_store(self):
         url = "http://127.0.0.1:8000/api/stores/"
         r = self.client.get(
-            url + str(self.store1.pk) + '/',
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
+            url + str(self.store1.pk) + "/",
+            HTTP_AUTHORIZATION="Bearer " + self.token.token,
         )
         self.assertEqual(200, r.status_code)
         self.assertEqual("Store 1", r.data["name"])
 
     def test_list_store(self):
         url = "http://127.0.0.1:8000/api/stores/"
-        r = self.client.get(
-            url,
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
-        )
+        r = self.client.get(url, HTTP_AUTHORIZATION="Bearer " + self.token.token)
         self.assertEqual(200, r.status_code)
         self.assertLessEqual(1, len(r.data))
 
@@ -217,7 +202,7 @@ class Test_StoreView(APITestCase):
         r = self.client.post(
             url,
             {"item_id": self.item1.pk},
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
+            HTTP_AUTHORIZATION="Bearer " + self.token.token,
         )
         self.assertEqual(200, r.status_code)
         self.assertEqual(2, len(self.store1.items.all()))
@@ -225,11 +210,13 @@ class Test_StoreView(APITestCase):
         self.assertEqual(1, len(self.store1.items.all()))
 
     def test_remove_item(self):
-        url = "http://127.0.0.1:8000/api/stores/" + str(self.store1.pk) + "/remove_item/"
+        url = (
+            "http://127.0.0.1:8000/api/stores/" + str(self.store1.pk) + "/remove_item/"
+        )
         r = self.client.post(
             url,
             {"item_id": self.item2.pk},
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
+            HTTP_AUTHORIZATION="Bearer " + self.token.token,
         )
         self.assertEqual(200, r.status_code)
         self.assertEqual(0, len(self.store1.items.all()))
@@ -241,7 +228,7 @@ class Test_StoreView(APITestCase):
         r = self.client.delete(
             url,
             {"item_id": self.item3.pk, "store_id": self.store1.pk},
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
+            HTTP_AUTHORIZATION="Bearer " + self.token.token,
         )
         self.assertEqual(200, r.status_code)
         self.assertEqual(0, Item.objects.filter(name="Item 3").count())
@@ -271,18 +258,15 @@ class Test_ItemView(APITestCase):
     def test_retrieve_item(self):
         url = "http://127.0.0.1:8000/api/items/"
         r = self.client.get(
-            url + str(self.item1.pk) + '/',
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
+            url + str(self.item1.pk) + "/",
+            HTTP_AUTHORIZATION="Bearer " + self.token.token,
         )
         self.assertEqual(200, r.status_code)
         self.assertEqual("Item 1", r.data["name"])
 
     def test_list_item(self):
         url = "http://127.0.0.1:8000/api/items/"
-        r = self.client.get(
-            url,
-            HTTP_AUTHORIZATION="Bearer " + self.token.token
-        )
+        r = self.client.get(url, HTTP_AUTHORIZATION="Bearer " + self.token.token)
         self.assertEqual(200, r.status_code)
         self.assertLessEqual(1, len(r.data))
 
@@ -312,14 +296,14 @@ class Test_OAuth(APITestCase):
     def test_OAuth_redirect(self, m):
         self.client.force_login(self.user)
         m.register_uri(
-            'POST',
-            'http://127.0.0.1:8000/o/token/',
-            text="{\"testdata\": \"hello world!\"}"
+            "POST",
+            "http://127.0.0.1:8000/o/token/",
+            text='{"testdata": "hello world!"}',
         )
-        url = 'http://127.0.0.1:8000/o/authorize/'
-        url += '?response_type=code'
-        url += '&client_id=' + self.application.client_id
-        url += '&redirect_uri=http://127.0.0.1:8000/authredirect/'
+        url = "http://127.0.0.1:8000/o/authorize/"
+        url += "?response_type=code"
+        url += "&client_id=" + self.application.client_id
+        url += "&redirect_uri=http://127.0.0.1:8000/authredirect/"
         r = self.client.get(url, follow=True)
         self.assertEqual(200, r.status_code)
         self.assertEqual(b'{"testdata":"hello world!"}', r.content)
