+++ backend/api/views.py
@@ -218,7 +218,7 @@ class RegisterUserViewSet(viewsets.ViewSet):
             return Response(data={"status": 0}, status=status.HTTP_400_BAD_REQUEST)
         else:
             application = Application.objects.get(
-                client_id=os.getenv("CLIENT_ID", 'defaultTestClientID')
+                client_id=os.getenv("CLIENT_ID", "defaultTestClientID")
             )
             TOKEN = "".join(random.choice(string.ascii_letters) for i in range(25))
             token = AccessToken.objects.create(
